updated the SurveyInfoPage component to use fetch, open the correct link, and display a loading message
